rename folders and hooks to dash
