+++ frontend/src/components/Results-View/Event-Footer-Actions/event-footer-actions.component.css
