+++ frontend/src/components/Results-View/Event-Footer-Actions/index.tsx
@@ -1,0 +1,1 @@
+export { EventFooterActions } from "./event-footer-actions.component";
