+++ frontend/src/components/Results-View/Item-Menu/Event-Header-Actions/event-header-actions.component.css
