+++ frontend/src/components/Results-View/Item-Menu/Event-Header-Actions/index.tsx
@@ -1,0 +1,1 @@
+export { EventHeaderActions } from "./event-header-actions.component";
