+++ frontend/src/components/Results-View/Item-Menu/Person-Card/Item-Row/index.tsx
@@ -1,0 +1,1 @@
+export { ItemRow } from "./item-row.component";
