+++ frontend/src/components/Results-View/Item-Menu/Person-Card/Person-Actions/index.tsx
@@ -1,0 +1,1 @@
+export { PersonActions } from "./person-actions.component";
