+++ frontend/src/components/Results-View/Item-Menu/Person-Card/Person-Actions/person-actions.component.tsx
@@ -1,4 +1,4 @@
-import "./personActions.component.css";
+import "./person-actions.component.css";
 import {
   addItem,
   deletePerson,
