+++ frontend/src/components/Results-View/Item-Menu/Person-Card/index.tsx
@@ -1,0 +1,1 @@
+export { PersonCard } from "./person-card.component";
