+++ frontend/src/components/Results-View/Item-Menu/Person-Card/person-card.component.css
