+++ frontend/src/components/Results-View/Item-Menu/Person-Card/person-card.component.tsx
@@ -3,9 +3,9 @@ import { useDispatch, useSelector } from "react-redux";
 import { didClickAway, didMount } from "../../../../hooks";
 import { AppStore, Item, Person } from "../../../../models";
 import { editPersonName } from "../../../../redux/calculation/calculation-actions";
-import { ItemRow } from "./ItemRow";
-import { PersonActions } from "./PersonActions";
-import "./personCard.component.css";
+import { ItemRow } from "./Item-Row";
+import { PersonActions } from "./Person-Actions";
+import "./person-card.component.css";
 import { validString } from "../../../../utilities/sanitize";
 import { dismissToast, errorToast } from "../../../shared/toasts/toasts";
 import { ERROR_INPUT_NAME } from "../../../../utilities/variables";
