+++ frontend/src/components/Results-View/Item-Menu/Totals-Menu/index.tsx
@@ -1,0 +1,1 @@
+export { TotalsMenu } from "./totals-menu.component";
