+++ frontend/src/components/Results-View/Item-Menu/Totals-Menu/totals-menu.component.css
