+++ frontend/src/components/Results-View/Item-Menu/index.tsx
@@ -1,0 +1,1 @@
+export { ItemMenu } from "./item-menu.component";
