+++ frontend/src/components/Results-View/index.tsx
@@ -1,0 +1,1 @@
+export { ResultsView } from "./results-view.component";
