+++ frontend/src/components/Results-View/results-view.component.css
