--- frontend/src/components/ResultsView/EventFooterActions/index.tsx
@@ -1,1 +1,0 @@
-export { EventFooterActions } from "./eventFooterActions.component";
