--- frontend/src/components/ResultsView/ItemMenu/EventHeaderActions/index.tsx
@@ -1,1 +1,0 @@
-export { EventHeaderActions } from "./eventHeaderActions.component";
