--- frontend/src/components/ResultsView/ItemMenu/PersonCard/ItemRow/index.tsx
@@ -1,1 +1,0 @@
-export { ItemRow } from "./itemRow.component";
