--- frontend/src/components/ResultsView/ItemMenu/PersonCard/PersonActions/index.tsx
@@ -1,1 +1,0 @@
-export { PersonActions } from "./personActions.component";
