--- frontend/src/components/ResultsView/ItemMenu/PersonCard/index.tsx
@@ -1,1 +1,0 @@
-export { PersonCard } from "./personCard.component";
