--- frontend/src/components/ResultsView/ItemMenu/TotalsMenu/index.tsx
@@ -1,1 +1,0 @@
-export { TotalsMenu } from "./totalsMenu.component";
