--- frontend/src/components/ResultsView/ItemMenu/index.tsx
@@ -1,1 +1,0 @@
-export { ItemMenu } from "./itemMenu.component";
