--- frontend/src/components/ResultsView/index.tsx
@@ -1,1 +1,0 @@
-export { ResultsView } from "./resultsView.component";
