+++ frontend/src/hooks/index.ts
@@ -1,0 +1,2 @@
+export { didMount } from "./did-Mount/did-mount.hook";
+export { didClickAway } from "./did-Click-Away/did-click-away.hook";
