--- frontend/src/hooks/index.tsx
@@ -1,2 +1,0 @@
-export { didMount } from "./didMount/didMount.hook";
-export { didClickAway } from "./didClickAway/didClickAway.hook";
